refactor(react): split FiberNodeButton into directory module (#145)
Break the ~1770-line packages/react/src/fiber-node-button.tsx into a
fiber-node-button/ directory split by responsibility, while keeping the
public API and runtime behavior unchanged.

New layout:
- fiber-node-button/types.ts             public + internal types, constants
- fiber-node-button/utils.ts             pure helpers (shorten, formatChannelBalance, ...)
- fiber-node-button/styles.ts            all inline CSSProperties
- fiber-node-button/use-panel-state.ts   panel state hook (useFiberNodeButtonPanelState)
- fiber-node-button/workbench-tab.tsx    Connection Prep / Open Channel / Payments
- fiber-node-button/channels-tab.tsx     Channel Summary / list / details
- fiber-node-button/diagnostics-tab.tsx  Connected Peers / Graph Snapshot
- fiber-node-button/panel.tsx            header, tab nav, status / error notices, force-close dialog
- fiber-node-button/index.tsx            slim FiberNodeButton wrapper + public type re-exports

The original path packages/react/src/fiber-node-button.tsx is kept as a
one-line shim that re-exports from ./fiber-node-button/index.js so that
existing imports (including the test file) continue to work.

No public API changes, no styling/copy/a11y changes.
Tests: 11/11 pass. Build + typecheck: green.

Closes #145
